sw/pc/requirements.txt
mnemonic==0.21
bip32==4.0
bech32==1.2.0
bitcoinlib==0.7.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
ecdsa==0.19.1
greenlet==3.1.1
idna==3.10
numpy==2.2.4
pycryptodome==3.22.0
pyperclip==1.9.0
pyserial==3.5
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
qrcode==8.1
readchar==4.2.1
